+++ backend/sivujetti/src/BaseAPI.php
@@ -8,17 +8,19 @@ use Pike\ArrayUtils;
  * Shared funtionality for UserThemeAPI, UserSiteAPI and UserPluginAPI.
  */
 abstract class BaseAPI {
+    public const ON_ROUTE_CONTROLLER_BEFORE_EXEC = "sivujetti:beforeExecRouteController";
+    public const ON_PAGE_BEFORE_RENDER = "sivujetti:onPageBeforeRender";
     /** @var string */
     protected string $namespace;
     /** @var \Sivujetti\SharedAPIContext */
-    protected SharedAPIContext $storage;
+    protected SharedAPIContext $apiCtx;
     /**
      * @param string $namespace
-     * @param \Sivujetti\SharedAPIContext $storage
+     * @param \Sivujetti\SharedAPIContext $apiCtx
      */
-    public function __construct(string $namespace, SharedAPIContext $storage) {
+    public function __construct(string $namespace, SharedAPIContext $apiCtx) {
         $this->namespace = $namespace;
-        $this->storage = $storage;
+        $this->apiCtx = $apiCtx;
     }
     /**
      * @param string $when e.g. "sivujetti:onPageBeforeRender"
@@ -26,14 +28,14 @@ abstract class BaseAPI {
      * @return int listener id
      */
     public function on(string $when, callable $thenDo): int {
-        return $this->storage->addEventListener($when, $thenDo);
+        return $this->apiCtx->addEventListener($when, $thenDo);
     }
     /**
      * @param string $url
      * @param array<string, string> $attrs = []
      */
     public function enqueueCssFile(string $url, array $attrs = []): void {
-        $this->storage->getDataHandle()->userDefinedAssets->css[] = (object) [
+        $this->apiCtx->userDefinedAssets->css[] = (object) [
             "url" => $url,
             "attrs" => $attrs,
         ];
@@ -43,14 +45,14 @@ abstract class BaseAPI {
      * @return bool
      */
     public function isCssFileEnqueued(string $url): bool {
-        return ArrayUtils::findIndexByKey($this->storage->getDataHandle()->userDefinedAssets->css, $url, "url") > -1;
+        return ArrayUtils::findIndexByKey($this->apiCtx->userDefinedAssets->css, $url, "url") > -1;
     }
     /**
      * @param string $url
      * @param array<string, string> $attrs = []
      */
     public function enqueueJsFile(string $url, array $attrs = []): void {
-        $this->storage->getDataHandle()->userDefinedAssets->js[] = (object) [
+        $this->apiCtx->userDefinedAssets->js[] = (object) [
             "url" => $url,
             "attrs" => $attrs,
         ];
@@ -60,6 +62,6 @@ abstract class BaseAPI {
      * @return bool
      */
     public function isJsFileEnqueued(string $url): bool {
-        return ArrayUtils::findIndexByKey($this->storage->getDataHandle()->userDefinedAssets->js, $url, "url") > -1;
+        return ArrayUtils::findIndexByKey($this->apiCtx->userDefinedAssets->js, $url, "url") > -1;
     }
 }