+++ backend/sivujetti/src/UserSite/UserSiteAPI.php
@@ -2,9 +2,10 @@
 
 namespace Sivujetti\UserSite;
 
-use Sivujetti\{BaseAPI, Template};
+use Sivujetti\{BaseAPI, SharedAPIContext, Template};
 use Sivujetti\BlockType\BlockTypeInterface;
 use Pike\{PikeException, Validation};
+use Sivujetti\UserPlugin\UserPluginInterface;
 
 /**
  * An API for BACKEND_PATH . site/Site.php classes.
@@ -20,8 +21,7 @@ class UserSiteAPI extends BaseAPI {
         if (($errors = Validation::makeValueValidator()->rule("identifier")->validate($name)))
             throw new PikeException(implode("\n", $errors),
                                     PikeException::BAD_INPUT);
-        $mut = $this->storage->getDataHandle();
-        $mut->blockTypes->{$name} = $instance;
+        $this->apiCtx->blockTypes->{$name} = $instance;
     }
     /**
      * Adds $url to a list of urls, that will be included at the end of
@@ -31,8 +31,7 @@ class UserSiteAPI extends BaseAPI {
      * @param string $url e.g. "my-site.bundle.js"
      */
     public function enqueueEditAppJsFile(string $url): void {
-        $mut = $this->storage->getDataHandle();
-        $mut->adminJsFiles[] = $url;
+        $this->apiCtx->adminJsFiles[] = $url;
     }
     /**
      * Adds $fileId to a list of names that can be used as $block->renderer.
@@ -41,9 +40,21 @@ class UserSiteAPI extends BaseAPI {
      * @see \Sivujetti\Template::completePath()
      */
     public function registerBlockRenderer(string $fileId): void {
-        // @allow \PikeException
+        // @allow \Pike\PikeException
         Template::completePath($fileId, allowSubFolders: true);
-        $mut = $this->storage->getDataHandle();
-        $mut->validBlockRenderers[] = str_contains($fileId, ":") ? $fileId : "site:{$fileId}";
+        $this->apiCtx->validBlockRenderers[] = str_contains($fileId, ":") ? $fileId : "site:{$fileId}";
+    }
+    /**
+     * @param string $name
+     * @return \Sivujetti\UserPlugin\UserPluginInterface|null
+     * @throws \Pike\PikeException
+     */
+    public function getPlugin(string $name): ?UserPluginInterface {
+        if ($this->apiCtx->getAppPhase() < SharedAPIContext::PHASE_READY_TO_EXECUTE_ROUTE_CONTROLLER)
+            throw new PikeException("You should call \$api->getPlugin() inside \$api->on(\$api::ON_ROUTE_" .
+                                    "CONTROLLER_BEFORE_EXEC) or \$api->on(\$api::ON_PAGE_BEFORE_RENDER) " .
+                                    "to ensure they're all loaded",
+                                    PikeException::ERROR_EXCEPTION);
+        return $this->apiCtx->userPlugins[$name] ?? null;
     }
 }